+++ web-share/src/svelte/components/NavbarButton.svelte
